change nap(), signal_timer_t to use int64_t for millisecond delays

+++ src/arch/timing.cc
@@ -9,21 +9,21 @@
 
 // nap()
 
-void nap(int ms) THROWS_NOTHING {
+void nap(int64_t ms) THROWS_NOTHING {
     if (ms > 0) {
         signal_timer_t timer(ms);
         timer.wait_lazily_ordered();
     }
 }
 
-void nap(int ms, signal_t *interruptor) THROWS_ONLY(interrupted_exc_t) {
+void nap(int64_t ms, signal_t *interruptor) THROWS_ONLY(interrupted_exc_t) {
     signal_timer_t timer(ms);
     wait_interruptible(&timer, interruptor);
 }
 
 // signal_timer_t
 
-signal_timer_t::signal_timer_t(int ms) {
+signal_timer_t::signal_timer_t(int64_t ms) {
     rassert(ms >= 0);
     timer = fire_timer_once(ms, &signal_timer_t::on_timer_ring, this);
 }
@@ -40,7 +40,7 @@ void signal_timer_t::on_timer_ring(void *v_timer) {
 
 // repeating_timer_t
 
-repeating_timer_t::repeating_timer_t(int frequency_ms, repeating_timer_callback_t *_ringee) :
+repeating_timer_t::repeating_timer_t(int64_t frequency_ms, repeating_timer_callback_t *_ringee) :
     ringee(_ringee) {
     rassert(frequency_ms > 0);
     timer = add_timer(frequency_ms, &repeating_timer_t::on_timer_ring, this);
